requirements.txt
asgiref==3.5.2
attrs==22.1.0
backports.zoneinfo==0.2.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
defusedxml==0.7.1
Django==4.1.3
django-cors-headers==3.13.0
django-filter==22.1
django-multiselectfield==0.1.12
django-phonenumber-field==7.0.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.4
flake8==6.0.0
idna==3.4
importlib-resources==5.10.1
inflection==0.5.1
isodate==0.6.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
jsonschema-spec==0.1.2
lazy-object-proxy==1.8.0
MarkupSafe==2.1.1
mccabe==0.7.0
more-itertools==9.0.0
oauthlib==3.2.2
openapi-schema-validator==0.3.4
openapi-spec-validator==0.5.1
packaging==21.3
parse==1.19.0
pathable==0.4.3
phonenumbers==8.13.0
pkgutil_resolve_name==1.3.10
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.5
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.11.0
